feat: enhance VehiculoForm to handle lockedModeloId and confirm mileage reduction
- Added lockedModeloId prop to VehiculoForm to prevent changing the model when creating from a specific model.
- Implemented a confirmation modal for reducing mileage, allowing users to confirm if they want to decrease the recorded mileage.
- Updated validation logic to ensure mileage can only be decreased during editing.
- Adjusted searchable and disabled states for the model selection based on lockedModeloId.

feat: update useAgendasMantenimiento and useMantenimientos hooks to invalidate vehicle queries

- Invalidate vehicle queries when completing an agenda or registering maintenance to ensure the odometer reflects the latest data.

feat: introduce useRequerimientoCategorias hook for fetching requirement categories

- Added a new hook to fetch existing requirement categories for use in forms, enhancing user experience by allowing category creation.

fix: correct terminology from "Vehículo unitario" to "Vehículo utilitario" across various components and hooks

- Updated labels and comments to reflect the correct terminology for utility vehicles.

feat: implement pieza handling in MantenimientoForm

- Added functionality to capture and manage pieces used during maintenance, including validation and total cost calculation.
- Enhanced the MantenimientoForm to allow adding, editing, and removing pieces, ensuring accurate inventory management.

feat: create new API endpoint for fetching requirement categories

- Implemented a new API function to retrieve requirement categories, improving the backend support for the new hook.

+++ api/src/functions/agenda-mantenimiento-completar.ts
@@ -9,7 +9,7 @@ const Schema = z.object({
   fecha:             z.string().date(),
   tipo:              z.string().max(80).trim().nullable().optional(),
   tecnico:           z.string().max(120).trim().nullable().optional(),
-  costo:             z.coerce.number().int().min(0).default(0),
+  costo:             z.coerce.number().min(0).default(0),
   km_actual:         z.coerce.number().int().min(0).default(0),
   observaciones:     z.string().trim().nullable().optional(),
   requerimiento_ids: z.array(z.number().int().positive()).min(1, 'Selecciona al menos un requerimiento'),
